refactor: 검증 패키지 분리
